Revert "moved utils and pages folders out of component folder, redid routing …"
