+++ client/src/components/pages/event/Event.js
@@ -1,6 +1,6 @@
 import React, { Component } from "react";
-import CommentBox from "../components/assets/CommentBox";
-import EventInfo from "../components/assets/EventInfo";
+import CommentBox from "../../assets/CommentBox";
+import EventInfo from "../../assets/EventInfo";
 
 
 class Event extends Component {
