+++ client/src/components/pages/login-signup/form-containers/SignupForm.js
@@ -2,13 +2,16 @@ import React, { Component } from "react";
 
 import API from "../../../utils/API";
 
-import { Input } from "../../../components/assets/form/Input";
-import { FormBtn } from "../../../components/assets/form/FormBtn";
+import { Input } from "../../../assets/form/Input";
+import { FormBtn } from "../../../assets/form/FormBtn";
 
 
 class Login extends Component {
 
     state = {
+        emailAddress: "",
+        firstName: "",
+        lastName: "",
         username: "",
         password: ""
     };
@@ -24,10 +27,16 @@ class Login extends Component {
 
     handleFormSubmit = event => {
         event.preventDefault();
-        if (this.state.username && 
+        if (this.state.emailAddress && 
+            this.state.firstName && 
+            this.state.lastName && 
+            this.state.username && 
             this.state.password
             ) {
-        API.login({
+        API.createUser({
+            emailAddress: this.state.emailAddress,
+            firstName: this.state.firstName,
+            lastName: this.state.lastName,
             username: this.state.username,
             password: this.state.password
         })
@@ -42,6 +51,30 @@ class Login extends Component {
     render() {
         return (
             <form>
+
+                <Input
+                    label={"Email Address:"}
+                    name="emailAddress"
+                    value={this.state.password}
+                    onChange={this.handleInputChange}
+                    placeholder="Password"
+                />
+
+                <Input
+                    label={"First Name:"}
+                    name="firstName"
+                    value={this.state.password}
+                    onChange={this.handleInputChange}
+                    placeholder="Password"
+                />
+
+                <Input
+                    label={"Last Name:"}
+                    name="lastName"
+                    value={this.state.password}
+                    onChange={this.handleInputChange}
+                    placeholder="Password"
+                />
 
                 <Input
                     label={"Username:"}
